+++ src/content/solanaexpl/feat-scripts/index.ts
@@ -1,3 +1,5 @@
 export { default as renderFundFlowButton } from './fund-flow'
 export { default as renderMainAddressLabel } from './main-address-label'
 export { default as renderTxPageAddressLabels } from './tx-address-labels'
+export { default as renderAlternativeParsers } from './quick2parsers'
+export { default as renderTransactionHashPhalconLink } from './transaction-hash-phalcon-link'