+++ src/content/solanafm/page-scripts/address.tsx
@@ -7,28 +7,22 @@ import {
   GET_SOLANAFM_ACCOUNT_INFO,
   GET_SOLANAFM_ACCOUNT_TRANSFERS
 } from '@common/constants'
+import { pickSolanaAddress } from '@common/utils'
 
 import {
   renderFundFlowButton,
   renderMainAddressLabel,
-  renderEnhancedLabels
+  renderEnhancedLabels,
+  renderTransactionHashPhalconLink
 } from '../feat-scripts'
 import { lazyLoad } from '../helper'
 
 const createAccountBox = () => {
   if ($('#md-account-box').length) return Promise.reject()
-  const isWalletAccount =
-    $(
-      '#__next > main > div:nth-of-type(2) > div:nth-of-type(1) > div > div > div:nth-of-type(1) > div > div > div > div > div:nth-of-type(1) > div > div > h1'
-    )
-      .text()
-      .indexOf('Wallet Account') !== -1
   const container = $(
-    isWalletAccount
-      ? '#__next > main > div:nth-of-type(2) > div:nth-of-type(1) > div > div > div:nth-of-type(1) > div > div:nth-of-type(1) > div > div > div:last-child > div'
-      : '#__next > main > div:nth-of-type(2) > div:nth-of-type(1) > div > div > div:nth-of-type(1) > div > div:nth-of-type(1) > div > div > div:last-child > div > div:nth-of-type(2) > div:nth-of-type(1)'
-  )
-  const mainAddress = container.find('>span:first-child').text().trim()
+    '#__next > main > div:nth-of-type(2) > div:nth-of-type(1) > div > div > div:nth-of-type(1) div.text-ellipsis'
+  ).parent()
+  const mainAddress = pickSolanaAddress(window.location.href)
   const accountBox = $(
     `<div id="md-account-box" class="flex items-center flex-wrap gap-6 pl-5" data-address="${mainAddress}"></div>`
   )
@@ -40,7 +34,20 @@ const createAccountBox = () => {
 }
 
 const initAddressPageScript = async () => {
-  const { fundFlow, enhancedLabels } = await store.get('options')
+  const { fundFlow, enhancedLabels, quick2Parsers } = await store.get('options')
+
+  lazyLoad(() => {
+    if (quick2Parsers) {
+      renderTransactionHashPhalconLink()
+      $(
+        '#__next > main > div:nth-of-type(2) > div:nth-of-type(1) > div > div > div:nth-of-type(2) > div:nth-of-type(2) > div:nth-of-type(2) > div > div > div > div:nth-of-type(2)> div:nth-of-type(2) button'
+      ).click(function () {
+        setTimeout(() => {
+          renderTransactionHashPhalconLink()
+        })
+      })
+    }
+  }, '.animate-pulse')
 
   browser.runtime.onMessage.addListener((message, _sender, sendResponse) => {
     if (message === GET_SOLANAFM_ACCOUNT_INFO) {