+++ src/content/solscan/helper.ts
@@ -3,13 +3,14 @@ import $ from 'jquery'
 export const lazyLoad = (
   callback: () => void,
   inspector: string,
+  isNegative = true,
   maxRetries = 60
 ) => {
-  const loading = !$(inspector).length
+  const loading = isNegative ? !$(inspector).length : !!$(inspector).length
   if (loading) {
     if (maxRetries > 0) {
       setTimeout(() => {
-        lazyLoad(callback, inspector, maxRetries - 1)
+        lazyLoad(callback, inspector, isNegative, maxRetries - 1)
       }, 500)
     }
   } else {