+++ src/content/solscan/page-scripts/account.tsx
@@ -6,12 +6,13 @@ import { GET_SOLSCAN_ACCOUNT_TAB_DATA, SOLSCAN_PAGES } from '@common/constants'
 import {
   renderFundFlowButton,
   renderMainAddressLabel,
-  renderEnhancedLabels
+  renderEnhancedLabels,
+  renderTransactionHashPhalconLink
 } from '../feat-scripts'
 import { lazyLoad } from '../helper'
 
 const initAccountPageScript = async () => {
-  const { fundFlow, enhancedLabels } = await store.get('options')
+  const { fundFlow, enhancedLabels, quick2Parsers } = await store.get('options')
 
   lazyLoad(() => {
     if (fundFlow) renderFundFlowButton(SOLSCAN_PAGES.ACCOUNT.name)
@@ -21,10 +22,19 @@ const initAccountPageScript = async () => {
     }
   }, '#account-tabs')
 
+  lazyLoad(
+    () => {
+      if (quick2Parsers) renderTransactionHashPhalconLink()
+    },
+    '.animate-pulse',
+    false
+  )
+
   browser.runtime.onMessage.addListener((message, _sender, sendResponse) => {
     if (message === GET_SOLSCAN_ACCOUNT_TAB_DATA) {
       requestIdleCallback(() => {
         if (enhancedLabels) renderEnhancedLabels()
+        if (quick2Parsers) renderTransactionHashPhalconLink()
       })
       sendResponse()
     }