+++ src/content/solscan/page-scripts/index.ts
@@ -1,3 +1,4 @@
 export { default as initAccountPageScript } from './account'
 export { default as initTxPageScript } from './tx'
 export { default as initTokenPageScript } from './token'
+export { default as initBlockPageScript } from './block'