+++ src/content/solscan/page-scripts/tx.tsx
@@ -3,16 +3,25 @@ import browser from 'webextension-polyfill'
 import { store } from '@src/store'
 import { GET_SOLSCAN_TRANSACTION } from '@common/constants'
 
-import { renderTxPageAddressLabels } from '../feat-scripts'
+import {
+  renderTxPageAddressLabels,
+  renderAlternativeParsers
+} from '../feat-scripts'
+import { lazyLoad } from '../helper'
 
 const initTxPageScript = async () => {
-  const { enhancedLabels } = await store.get('options')
+  const { enhancedLabels, quick2Parsers } = await store.get('options')
 
   browser.runtime.onMessage.addListener((message, _sender, sendResponse) => {
     if (message === GET_SOLSCAN_TRANSACTION) {
-      requestIdleCallback(() => {
-        if (enhancedLabels) renderTxPageAddressLabels()
-      })
+      lazyLoad(
+        () => {
+          if (enhancedLabels) renderTxPageAddressLabels()
+          if (quick2Parsers) renderAlternativeParsers()
+        },
+        '.animate-pulse',
+        false
+      )
       sendResponse()
     }
   })